refactor(ds): lift the dialog's Tab handling out of its key listener
The gates cap cognitive complexity at 15 and the combined Escape + Tab
listener reached 19. The wrap logic is its own decision anyway: what a Tab
means at the edge of a dialog does not depend on how the key arrived.
